requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
bleach==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.0
cycler==0.12.1
eval_type_backport==0.2.2
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
kaggle==1.7.4.5
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
simsimd==6.5.3
six==1.17.0
stringzilla==4.2.1
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
timm==1.0.20
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
webencodings==0.5.1
zipp==3.23.0
gradio==4.7.1
